revamped ch3 with CUDA

+++ vec3.h
@@ -16,42 +16,42 @@ class vec3 {
     public:
         __host__ __device__ vec3() : e{0,0,0} {}                                            // initialize 3D vector at origin
 
-        __host__ __device__ vec3(double e0, double e1, double e2) : e{e0, e1, e2} {}        // initialize 3D vector with custom coordinates
+        __host__ __device__ vec3(float e0, float e1, float e2) : e{e0, e1, e2} {}        // initialize 3D vector with custom coordinates
 
-        __host__ __device__ double x() const { return e[0]; }
-        __host__ __device__ double y() const { return e[1]; }
-        __host__ __device__ double z() const { return e[2]; }
+        __host__ __device__ float x() const { return e[0]; }
+        __host__ __device__ float y() const { return e[1]; }
+        __host__ __device__ float z() const { return e[2]; }
 
     // operator overloading
     public:
 
         __host__ __device__ inline const vec3& operator+() const { return *this; }                  // addition
         __host__ __device__ inline vec3 operator-() const { return vec3(-e[0], -e[1], -e[2]); }     // inverse
-        __host__ __device__ inline double operator[](int i) const { return e[i]; }                  // const index
-        __host__ __device__ inline double& operator[](int i) { return e[i]; };                      // mutatable index
+        __host__ __device__ inline float operator[](int i) const { return e[i]; }                  // const index
+        __host__ __device__ inline float& operator[](int i) { return e[i]; };                      // mutatable index
 
         __host__ __device__ inline vec3& operator+=(const vec3 &v2);                                // self addition
         __host__ __device__ inline vec3& operator-=(const vec3 &v2);                                // self subtraction
         __host__ __device__ inline vec3& operator*=(const vec3 &v2);                                
         __host__ __device__ inline vec3& operator/=(const vec3 &v2);                                // 
-        __host__ __device__ inline vec3& operator*=(const double t);                                // self scalar multiplcation
-        __host__ __device__ inline vec3& operator/=(const double t);                                // self scalar division
+        __host__ __device__ inline vec3& operator*=(const float t);                                // self scalar multiplcation
+        __host__ __device__ inline vec3& operator/=(const float t);                                // self scalar division
 
     // methods
     public:
         /// return the length of a 3D vector.
-        __host__ __device__ double length() const {
+        __host__ __device__ float length() const {
             return sqrt(length_squared());
         }
 
         /// return the dot product of a 3D vector.
-        __host__ __device__ double length_squared() const {
+        __host__ __device__ float length_squared() const {
             return e[0]*e[0] + e[1]*e[1] + e[2]*e[2];
         }
 
     // attributes
     public:
-        double e[3];
+        float e[3];
 };
 
 
@@ -106,15 +106,15 @@ __host__ __device__ inline vec3 operator/(const vec3 &v1, const vec3 &v2) {
 __host__ __device__ inline vec3 operator*(const vec3 &v, float t) {
     return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
-__host__ __device__ inline vec3 operator*(double t, const vec3 &v) {
+__host__ __device__ inline vec3 operator*(float t, const vec3 &v) {
     return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
-__host__ __device__ inline vec3 operator/(vec3 v, double t) {
+__host__ __device__ inline vec3 operator/(vec3 v, float t) {
     return (1/t) * v;
 }
 
-__host__ __device__ inline double dot(const vec3 &u, const vec3 &v) {
+__host__ __device__ inline float dot(const vec3 &u, const vec3 &v) {
     return u.e[0] * v.e[0]
          + u.e[1] * v.e[1]
          + u.e[2] * v.e[2];
@@ -154,15 +154,15 @@ __host__ __device__ inline vec3& vec3::operator-=(const vec3& v) {
     return *this;
 }
 
-__host__ __device__ inline vec3& vec3::operator*=(const double t) {
+__host__ __device__ inline vec3& vec3::operator*=(const float t) {
     e[0] *= t;
     e[1] *= t;
     e[2] *= t;
     return *this;
 }
 
-__host__ __device__ inline vec3& vec3::operator/=(const double t) {
-    double k = 1.0/t;
+__host__ __device__ inline vec3& vec3::operator/=(const float t) {
+    float k = 1.0/t;
 
     e[0] *= k;
     e[1] *= k;
